requirements-dev.txt
-r requirements.txt

pytest-asyncio==0.14.0
respx==0.16.3

isort==5.7.0
mypy==0.812
safety==1.10.3
cognitive-complexity==1.2.0

mccabe==0.6.1
flake8==3.9.0
flake8-blind-except==0.2.0
flake8-broken-line==0.3.0
flake8-bugbear==21.3.2
flake8-builtins==1.5.3
flake8-class-attributes-order==0.1.2
flake8-cognitive-complexity==0.1.0
flake8-commas==2.0.0
flake8-comprehensions==3.4.0
flake8-debugger==4.0.0
flake8-eradicate==1.0.0
flake8-functions==0.0.5
flake8-isort==4.0.0
flake8-mock==0.3
flake8-mutable==1.2.0
flake8-print==4.0.0
flake8-pytest==1.3
flake8-pytest-style==1.4.0
flake8-quotes==3.2.0
flake8-string-format==0.3.0
flake8-variables-names==0.0.4
